% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NFL_QBR_by_year}
\alias{NFL_QBR_by_year}
\title{NFL QBR by Year}
\format{
\subsection{\code{NFL_QBR_by_year}}{

A data frame with 1,386 rows and 4 columns:
\describe{
\item{Player}{Player name}
\item{Season}{Year}
\item{Att}{Total attempts}
\item{QBR}{Total Quarterback Rating}
}
}
}
\source{
\url{https://www.pro-football-reference.com/years/2024/passing.htm}
}
\usage{
NFL_QBR_by_year
}
\description{
NFL quarterback performance by year according to ESPN's Total Quarterback Rating, 2006-2024.
}
\keyword{datasets}
