% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quarterback}
\alias{quarterback}
\title{Historical Quarterback Prospects}
\format{
\subsection{\code{quarterback}}{

A data frame with 5,701 rows and 27 columns:
\describe{
\item{player_name}{Player name}
\item{ncaa_years_career}{Number of seasons played in college career}
\item{ncaa_games_career}{Total number of games in college career}
\item{ncaa_yds_per_att_career}{Passing yards per attempt in college career}
\item{ncaa_games_per_year}{Games per year in college career}
\item{ncaa_att_per_year}{Passing attempts per year in college career}
\item{ncaa_cmp_per_year}{Passing completions per year in college career}
\item{ncaa_yds_per_year}{Passing yards completions per year in college career}
\item{ncaa_td_per_year}{Passing touchdowns per year in college career}
\item{ncaa_int_per_year}{Passing interceptions per year in college career}
\item{ncaa_rush_att_per_year}{Rushing attempts per year in college career}
\item{ncaa_rush_yds_per_year}{Rusing yards per year in college career}
\item{ncaa_rush_td_per_year}{Rushing touchdowns per year in college career}
\item{ncaa_season_last}{Year of final college season}
\item{ncaa_conference_last}{Conference for final college season}
\item{ncaa_team_last}{Team for final college season}
\item{ncaa_sos_last}{Strength of schedule for final college season}
\item{ncaa_games_last}{Games played in final college season}
\item{ncaa_yds_per_att_last}{Passing yards per attempt in final college season}
\item{ncaa_passer_rating_last}{Passer rating in final college season}
\item{ncaa_all_america}{Numer of All-America recognitions in college career}
\item{ncaa_heisman}{Number of Heisman awards in college career}
\item{ncaa_heisman_last}{Heisman voting finish in final college season}
\item{nfl_years}{Number of seasons played in NFL career}
\item{nfl_att_career}{Total passing attempts in NFL career}
\item{nfl_att_per_year}{Passing attempts per year in NFL career}
\item{nfl_qbr_per_year}{Average QBR (across years) in NFL career}
}
}
}
\source{
\url{https://www.sports-reference.com/cfb/years/2024-passing.html}
}
\usage{
quarterback
}
\description{
College careeer statistics and career NFL performance (if applicable) for historical college
quarterback prospects from 2001 through 2024. Only NFL performance through 2024 is included.
}
\keyword{datasets}
